--- src/tqec/circuit/detector/__init__.py
@@ -1,4 +1,0 @@
-from tqec.circuit.detector.detector_util import (
-    PauliString,
-    iter_stim_circuit_by_moments,
-)
