+++ src/tqec/circuit/detector/pauli_test.py
@@ -1,11 +1,6 @@
 import stim
 
-from tqec.circuit.detector.detector_util import (
-    Fragment,
-    FragmentLoop,
-    PauliString,
-    split_stim_circuit_into_fragments,
-)
+from tqec.circuit.detector.pauli import PauliString
 
 
 def test_pauli_string_construction():
@@ -60,34 +55,3 @@ def test_pauli_string_collapse_by():
     c = PauliString({1: "X"})
     assert a.collapse_by([b]) == PauliString({0: "I", 1: "Y"})
     assert a.collapse_by([c]) == PauliString({0: "X", 1: "I"})
-
-
-def test_split_stim_circuit_into_fragments():
-    d = 5
-    surface_code_circuit_with_mr = stim.Circuit.generated(
-        code_task="surface_code:rotated_memory_x",
-        distance=d,
-        rounds=d,
-        after_clifford_depolarization=0.001,
-        after_reset_flip_probability=0.01,
-        before_measure_flip_probability=0.01,
-        before_round_data_depolarization=0.005,
-    )
-    fragments = split_stim_circuit_into_fragments(surface_code_circuit_with_mr)
-    assert len(fragments) == 3
-    f1, f2, f3 = fragments
-    assert isinstance(f1, Fragment)
-    assert isinstance(f2, FragmentLoop)
-    assert isinstance(f3, Fragment)
-    assert f1.circuit.num_measurements == d**2 - 1
-    assert f1.circuit.num_ticks == 7
-    assert len(f1.end_stabilizer_sources) == 2 * d**2 - 1
-    assert len(f1.begin_stabilizer_sources) == d**2 - 1
-    assert len(f1.sources_for_next_fragment) == d**2 - 1
-
-    assert f2.repetitions == d - 1
-    assert len(f2.fragments) == 1
-    assert isinstance(f2.fragments[0], Fragment)
-    assert f2.fragments[0].circuit[-38:-26] == f1.circuit[-25:-13]
-
-    assert f3.circuit.num_measurements == d**2
